fix(acl): split top-menu upgrade helpers for PHPCS side effects
Keep function declarations out of the upgrade runner so PSR-1
SideEffects passes. Refs #14498.
